requirements.txt
flask
flask-cors
pymongo
marshmallow
python-dotenv